+++ packages/asc-ui/src/components/ErrorMessage/index.ts
@@ -1,0 +1,2 @@
+export { default, ErrorMessageProps } from './ErrorMessage'
+export { default as ErrorMessageStyle } from './ErrorMessageStyle'